requirements.txt
streamlit
google-cloud-bigquery==3.26.0
google-auth-httplib2==0.2.0
pandas
plotly.express